Add support for setting a channel binding token
A channel binding token is the link between a GSSAPI exchange and
the transport connection over which it's performed. Active Directory
can enforce channel bindings on TLS connections as a mitigation
against authentication relaying attacks.

This patch adds a method to set a channel binding token on a
ClientCtx, which will be consumed with the next step() call
on the context.